still working on area of focus bug with each character as an element
