+++ data/tos.txt
@@ -1,2 +1,2 @@
 Netflix Terms of Use
-Netflix provides a personalized subscription service that allows our members to access entertainment content 
+Netflix provides a personalized subscription service that allows our members to access entertainment content
