+++ fixtures/vcr_cassettes/synopsis.yaml
@@ -1,0 +1,61 @@
+interactions:
+- request:
+    body: '{"user": 1, "areas_of_focus": ["mandatory binding arbitration", "recurring
+      payments"], "tos": "Netflix Terms of Use\nNetflix provides a personalized subscription
+      service that allows our members to access entertainment content "}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '228'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - python-requests/2.29.0
+    method: POST
+    uri: http://localhost:8000/api/queries
+  response:
+    body:
+      string: '{"id":20,"response":"[\"\\nAREA OF FOCUS: mandatory binding arbitration
+        \\nREQUEST 1: Netflix''s Terms of Use states that all disputes between the
+        company and its members must be resolved through binding arbitration. This
+        means that any disputes must be settled through a third-party arbitrator,
+        rather than through a court of law. \\nREQUEST 2: The Terms of Use outlines
+        the steps that members can take to protect themselves in the event of a dispute.
+        These steps include filing a claim with the arbitrator, providing evidence
+        to support the claim, and attending the arbitration hearing. \\nREQUEST 3:
+        8.\", ''\\nAREA OF FOCUS: recurring payments \\nREQUEST 1: Netflix allows
+        members to pay for their subscription on a monthly basis. The terms of service
+        outline that members will be charged on a recurring basis until they cancel
+        their subscription. \\nREQUEST 2: The terms of service outline that members
+        can cancel their subscription at any time. Members can also change their payment
+        method or update their billing information. \\nREQUEST 3: 8'']","query":36}'
+    headers:
+      Allow:
+      - GET, POST, HEAD, OPTIONS
+      Content-Length:
+      - '1061'
+      Content-Type:
+      - application/json
+      Cross-Origin-Opener-Policy:
+      - same-origin
+      Date:
+      - Tue, 16 May 2023 11:39:31 GMT
+      Referrer-Policy:
+      - same-origin
+      Server:
+      - WSGIServer/0.2 CPython/3.11.3
+      Vary:
+      - Cookie
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - DENY
+    status:
+      code: 201
+      message: Created
+version: 1
