+++ requirements.txt
@@ -41,6 +41,7 @@ pyasn1==0.5.0
 pyasn1-modules==0.3.0
 pydantic==1.10.7
 pyparsing==3.0.9
+pytest==7.3.1
 pytest-django==4.5.2
 python-dateutil==2.8.2
 python-decouple==3.8
@@ -48,6 +49,7 @@ pytz==2023.3
 PyYAML==6.0
 regex==2023.5.5
 requests==2.29.0
+requests-mock==1.10.0
 requests-oauthlib==1.3.1
 rsa==4.9
 six==1.16.0
@@ -61,4 +63,6 @@ typing_extensions==4.5.0
 tzdata==2023.3
 uritemplate==4.1.1
 urllib3==1.26.15
+vcrpy==4.2.1
+wrapt==1.15.0
 yarl==1.9.2
